requirements.txt
fastapi
bytetracker
uvicorn
pillow
python-dotenv
boto3
paddleocr
paddlepaddle
ultralytics
onnx
tensorrt